ported db usage
