+++ src/Models/Category.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * @property string $guid primary key
+ * @property string $title
+ */
+class Category extends Model
+{
+    
+}
